Upgraded Pawn compiler to upstream SVN r37.

+++ Compiler/source/compiler/memfile.h
@@ -2,7 +2,7 @@
  *  memory.
  *
  *  Copyright (c) faluco / http://www.amxmodx.org/, 2006
- *  Version: $Id: memfile.h 4240 2010-04-06 15:55:46Z thiadmer $
+ *  Version: $Id: memfile.h 4611 2011-12-05 17:46:53Z thiadmer $
  */
 
 #ifndef _INCLUDE_MEMFILE_H
@@ -12,8 +12,8 @@ typedef struct memfile_s
 {
 	char *name;
 	char *base;
-	long offs;
-	long usedoffs;
+	size_t offs;
+	size_t usedoffs;
 	size_t size;
 } memfile_t;
 
@@ -22,6 +22,6 @@ void memfile_destroy(memfile_t *mf);
 void memfile_seek(memfile_t *mf, long seek);
 int memfile_write(memfile_t *mf, const void *buffer, size_t size);
 size_t memfile_read(memfile_t *mf, void *buffer, size_t maxsize);
-long memfile_tell(const memfile_t *mf);
+size_t memfile_tell(const memfile_t *mf);
 
 #endif //_INCLUDE_MEMFILE_H
