+++ Compiler/source/compiler/sclist.c
@@ -20,7 +20,7 @@
  *  License for the specific language governing permissions and limitations
  *  under the License.
  *
- *  Version: $Id: sclist.c 4548 2011-08-01 09:35:40Z thiadmer $
+ *  Version: $Id: sclist.c 4611 2011-12-05 17:46:53Z thiadmer $
  */
 #include <assert.h>
 #include <limits.h>
@@ -213,14 +213,14 @@ SC_FUNC stringpair *insert_alias(char *name,char *alias)
   assert(strlen(name)<=sNAMEMAX);
   assert(alias!=NULL);
   assert(strlen(alias)<=sNAMEMAX);
-  if ((cur=insert_stringpair(&alias_tab,name,alias,strlen(name)))==NULL)
+  if ((cur=insert_stringpair(&alias_tab,name,alias,(int)strlen(name)))==NULL)
     error(103);       /* insufficient memory (fatal error) */
   return cur;
 }
 
 SC_FUNC int lookup_alias(char *target,char *name)
 {
-  stringpair *cur=find_stringpair(alias_tab.next,name,strlen(name));
+  stringpair *cur=find_stringpair(alias_tab.next,name,(int)strlen(name));
   if (cur!=NULL) {
     assert(strlen(cur->second)<=sNAMEMAX);
     strcpy(target,cur->second);
